+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project changes will be shown here.
 - Issue 246 - Add descriptions to Report pages
 - Issue 237 - ServiceRequestReport export tweak
 - Issue 244 - Show ItemsNotYetReceivedEmail as a menu link
+- Issue 245 - Outstanding SRs email to display on screen as well
 ### Changed
 - Issue 233 - Amend Create Renewal Sales Order email destination
 - Issue 238 - SR Response and number amendment